refactor: type 분리 - #26
