+++ src/features/community/components/PostContent.tsx
@@ -1,8 +1,5 @@
 import styled from 'styled-components';
-
-interface PostContentPropTypes {
-  content: string;
-}
+import { PostContentPropTypes } from '../types';
 
 const PostContent = ({ content }: PostContentPropTypes) => {
   return (
